Amend add_schedule_service
Cases that are Offender SARs don't have to add retention_schedules
their linked cases on closure closure or when the initial script runs.

This is because for Branston cases Offender SARs only link to Offender
SAR complaints and not other Offender SARs.

The script was having an issue whereby it was updating linked cases for
Offender SARs, which meant that planned destruction date for the
associated complaints was being updated to match the complaints, when it
should have been the case that the Offender SAR complaint planned destruction
dates were mirrored in their associated Offender SAR cases.