CT-4219 Hide change links/buttons if case is readonly
This is a follow-up to PR #2032 where the backend side was introduced.

In this PR we hide from the UI functionality like change links, buttons, etc. that is no longer usable once a case has been anonymised and thus is in read only mode.

I might have missed some place but it easy to fix later on. And in any case even if the user tries to update something the backend side will stop them raising an exception.

Note: I've tried to piggy-back as much as possible on the already existing `allow_editing` partial propagation. However this affects mainly the Change links, not the buttons and some other actions, so I had to add some conditionals in some other places.