CT-4231 Change GDPR date filters from/to values
The from/to dates in the GDPR filters had an incorrect logic making them look for destruction dates in the past, instead of in the future (for example 1 month ago from today, instead of 1 month in the future from today).

Also amended the copy of the different options and support for i18n.

The new logic is as follow:

- Due - where `destruction date >= 01/01/2000` AND `destruction date <= TODAY`
- In 1 month - where `destruction date >= TODAY+1day` AND `destruction date <= TODAY+1month`
- In 2 months - where `destruction date >= TODAY+1month+1day` AND `destruction date <= TODAY+2months`
- In 3 months - where `destruction date >= TODAY+2months+1day` AND `destruction date <= TODAY+3months`
- In 4 months - where `destruction date >= TODAY+3months+1day` AND `destruction date <= TODAY+4months`