CT-4234 Add retention case notes on SAR closure (#2045)
When a case is closed, the system calculates and updates the case's retention schedule. It also looks for any linked cases and updates the destruction date on all those related cases.

Upon any of these updates, a case note will be added to any affected case, stating the retention schedule changes, but noting this was a "System update" as it was not a direct action of the user (the user just closed a case).

If this is for new retention schedules, state will always be `Not set` (the default), so we don't add this to the case note, we only add the destruction date.